Clean up TokenDetail
