+++ packages/wallet/src/App/index.tsx
@@ -7,7 +7,7 @@ import { pathLogin, pathOperationResult, pathTokens } from "./paths";
 import Login from "./routes/Login";
 import OperationResult from "./routes/OperationResult";
 import TokenDetail from "./routes/TokenDetail";
-import TokenList from "./routes/TokenList";
+import Tokens from "./routes/Tokens";
 
 function App(): JSX.Element {
   return (
@@ -19,7 +19,7 @@ function App(): JSX.Element {
               <Route exact path="/" component={Login} />
               <Route exact path={pathLogin} component={Login} />
               <ProtectedSwitch>
-                <Route exact path={pathTokens} component={TokenList} />
+                <Route exact path={pathTokens} component={Tokens} />
                 <Route exact path={`${pathTokens}/:tokenName`} component={TokenDetail} />
                 <Route exact path={pathOperationResult} component={OperationResult} />
               </ProtectedSwitch>
