+++ packages/wallet/src/App/routes/TokenDetail/FormSendTokens/FormSendTokens.less
@@ -1,0 +1,28 @@
+/* Mixins: .StandardPageLayout() */
+@import "~@cosmicdapp/design/theme/layout/mixins.less";
+
+.FormSendTokens {
+  --gap: var(--s0);
+
+  .formField {
+    display: flex;
+    align-items: baseline;
+
+    & > * + * {
+      margin-left: var(--s0);
+    }
+
+    span {
+      font-family: Iceland;
+      font-size: var(--s2);
+    }
+  }
+
+  .ant-form-item {
+    margin-bottom: 0;
+  }
+
+  button {
+    margin-top: var(--s2);
+  }
+}
