+++ packages/wallet/src/App/routes/TokenDetail/FormSendTokens/index.tsx
@@ -1,0 +1,64 @@
+import { Button, Typography } from "antd";
+import { Formik } from "formik";
+import { Form, FormItem, Input } from "formik-antd";
+import React from "react";
+import * as Yup from "yup";
+import Stack from "../../../../theme/layout/Stack";
+import { sendAddressValidationSchema } from "../../../forms/validationSchemas";
+import "./FormSendTokens.less";
+
+const { Text } = Typography;
+
+interface FormSendTokensProps {
+  readonly tokenName: string;
+  readonly tokenAmount: string;
+  readonly sendTokensAction: (values: any) => void;
+}
+
+function FormSendTokens({ tokenName, tokenAmount, sendTokensAction }: FormSendTokensProps): JSX.Element {
+  const sendAmountValidationSchema = Yup.object().shape({
+    amount: Yup.number()
+      .required("An amount is required")
+      .positive("Amount should be positive")
+      .max(parseFloat(tokenAmount), `Amount cannot be greater than ${tokenAmount}`),
+  });
+
+  const sendValidationSchema = sendAmountValidationSchema.concat(sendAddressValidationSchema);
+
+  return (
+    <Formik
+      initialValues={{ amount: "", address: "" }}
+      onSubmit={sendTokensAction}
+      validationSchema={sendValidationSchema}
+    >
+      {(formikProps) => (
+        <Form>
+          <Stack className="FormSendTokens">
+            <div className="formField">
+              <Text>Send</Text>
+              <FormItem name="amount">
+                <Input name="amount" placeholder="Enter amount" />
+              </FormItem>
+              <Text>{tokenName}</Text>
+            </div>
+            <div className="formField">
+              <Text>to</Text>
+              <FormItem name="address">
+                <Input name="address" placeholder="Enter address" />
+              </FormItem>
+            </div>
+            <Button
+              type="primary"
+              onClick={formikProps.submitForm}
+              disabled={!(formikProps.isValid && formikProps.dirty)}
+            >
+              Send
+            </Button>
+          </Stack>
+        </Form>
+      )}
+    </Formik>
+  );
+}
+
+export default FormSendTokens;
